Latest update (v7)
Should be fully localised now, for both English and French locales.

Changed the file encoding to ANSI-1252 (windows-1252) to facilitate script editing; the code page of the command prompt is then changed by the script to match that before runninng its code.

Script now has built-in support for the selection and merging of video-only and audio-only streams. Downloading a standalone stream (either video-only or audio-only) should also possible.